Make inspector usable in opt builds
Summary:
The Inspector assumed debugging was always present, making it rely on Hermes symbols that are not available in opt builds.

This change makes Hermes offer a no-op debugger in cases where the debugger is not included. This allows consumers (e.g. the Inspector) to use the same code path in debug and release builds.

## Sample Use

The Chrome Debug Server can be built with `xplat/mode/hermes/opt`, and profile local javascript:
https://pxl.cl/2c1vg

With some minor adjustments to the RN build (outside of the Inspector), FB4A can be built using `xplat/mode/hermes/opt` and profile the Marketplace app.
Via Chrome Dev Tools:
https://pxl.cl/2c1vm

and via Hermes Debugger (RN) in Flipper:
https://pxl.cl/2c1vt

Changelog: [Internal]

Reviewed By: jpporto

Differential Revision: D39120719

fbshipit-source-id: 2d34c52f7eac3e06154a8c85475439ba834721d0