Fix compilation for Windows ARM64 (#808)
Summary:
Currently code in `OSCompatWindows.cpp` causes a compilation error for ARM64 platform.
The issue is that the `__rdtsc()` intrinsic function is defined only for x86 and x64 CPUs and it is not available for ARM64.
The PR addresses this issue and also fixes incorrect `return` statement in the `#else` branch that is used for ARM64.

Pull Request resolved: https://github.com/facebook/hermes/pull/808

Test Plan: All unit tests are passing. No changes in code flow.

Reviewed By: mattbfb

Differential Revision: D39412176

Pulled By: jpporto

fbshipit-source-id: 4f0043638def03b64d78db9b28bff04ffedda789