Define SHLegacyValue
Summary:
SHlegacyValue is a C struct that defines the format of HV64. Move all
constants and type definitions from HermesValue.h to the new header.

HermesValue publicly inherits from SHLegacyValue, so they have a
compatible layout. This is intended to avoid UB when pointers are cast
between them.

For additional certainty, strict aliasing will be disabled in the
project.

Reviewed By: avp

Differential Revision: D39161676

fbshipit-source-id: 84c551fcf9953ac0fa2bb7397982ca3ed2520365