Adding internal fuzzilli function
Summary:
As specified in the comment above the function. Fuzzilli requires a "fuzzilli" function to run. This is used, among other things, to test crash detection (e.g. that an assert violation actually produces a crash).

In the past this function was defined in the REPRL used by fuzzilli to interface w/ Hermes. Though, this approach has some shortcoming (which I believe have caused a number of bugs to be missed few years ago when we first ran fuzzilli on Hermes): the REPRL is compiled separately by buck and the flags (e.g. to enable assert) are not forwarded to its dependencies (the VM).

Therefore, we should move this builtin inside Hermes VM. This will ensure that bugs in our build system for fuzzilli harnesses don't go undetected.

Otherwise, if something changes in the current way we propagate buck modes (T130118820), our harnesses will break w/o us knowing.

Reviewed By: jpporto

Differential Revision: D38905466

fbshipit-source-id: 71891818864f207e1d5ff7558885ee1dc083c82f