Use multiple cores when building with CMake
Summary:
When installing RNTester in React Native we build hermes from source. Currently, we are not asking our users to install ninja, so the build fallback uses CMake. We realized that we have very long builds and that is because we are not using multiple cores when building Hermes.

This diff tries to address that problem, using the same code we are using in other CI jobs

Differential Revision: D39538731

fbshipit-source-id: 41e5d7a24e5fa1a5ffca115d3f0a282bd474ea1a