Add object/array allocation with buffers
Summary:
The SerializedLiteralGenerator needs to be reusable by other backends,
so refactor it to take a couple `std::function`s to lookup the string
information.

The SerializedLiteralParser needs completely different logic for looking
up information in `SHUnit` compared to `RuntimeModule`,
so make a new `SHSerializedLiteralParser` for that purpose.
This is easy because `SerializedLiteralParser` already inherits from a
base class so only some small code needs copying.

Store the buffers and literal generator in a `SHLiteralBuffers` class,
and generate it much like the string table.

Reviewed By: tmikov

Differential Revision: D39485550

fbshipit-source-id: be267130513097a388ccbf0ce44513d60cc6313a