Move MemorySizeParser and RandomSeedParser to support
Summary:
Move the two parsers to Support so they can be used by consumers not
linking to ConsoleHost.

MemorySizeParser used to have a dependency on GCConfig for
`gcheapsize_t`, but didn't actually use it. Files in Support should not
depend on the GC, so I removed the dependency for now, but it is perhaps
not ideal.

Reviewed By: avp

Differential Revision: D39634643

fbshipit-source-id: da8d56e0c329160d935ff7c6dc90b9d10e9bd88a