produce a single hermesvm library
Summary:
The separate library per component strategy is straight-forward to create
and simplifies unit testing and simple tools. However it is very
inconvenient for users of the whole, since the names and paths of all
internal libraries are just implementation details. An embedder of
Hermes should ideally just need to link against `libhermes.a`.

This is what this commit does - it produces three standalone libraries:
- `libhermescompiler.a`
- `libhermesvm.a`
- `libhermesvmlean.a`

The `hermes`, `hermesc`, and `hvm` tools, as well as the JSI API, just
link to the corresponding library.

## Implementation Details

This is a rather big change, touching all CMakeLists.txt files.

It attempts to preserve the previous capabilities in a natural way (the
ability to just link against a library defined in a directory) while
generating the new "combined" libraries, hopefully without too many
hacks. It doesn't complete succeed to avoid hacks, but it passes
tests.

First, support for HERMES_LINK_COMPONENTS is removed, since it is
unnecessary and confusing - just an extra implicit way to add
dependencies to a target, as if we didn't have enough already.

The big change that enables the "amalgamation" of smaller libraries into
a large combined one, without hacks, is using CMake's "OBJECT"
libraries.

A CMake "OBJECT" library is an abstract collection of objects with
metadata, which acts in many ways similar to a regular library: it can
be linked against, and it propagates metadata transitively (defines,
include directories, etc).

One critical way in which it differs from normal libraries is that it
doesn't transitively propagate "OBJECT" library dependencies. In other
words, if OBJECT library B depends on OBJECT library A, and we link
a binary against B, we will get link errors for A's symbols.

So, it is impossible to just replace normal libraries with object
libraries, we need to be aware of the distinction.

### add_hermes_library() changes

We have extended `add_hermes_library()` to support this new notion of
OBJECT libraries, by recognizing the STATIC, SHARED or OBJECT keyword,
defaulting to OBJECT. If STATIC or SHARED is used, it works as it did
previously. If no keyword, or OBJECT, is specified it creates the new
type of library.

The dependencies that can be supploed to `add_hermes_library()` are also
extended: "normal" libraries are specified as before with `LINK_LIBS`,
but object library dependencies are specified with `LINK_OBJLIBS`. Since
we have converted most libraries to OBJECT, `LINK_OBJLIBS` is now more
common.

Example:
```
add_hermes_library(foo OBJECT f1.c
        LINK_OBJLIBS dep1 dep2
        LINK_LIBS pthread)
target_compile_definitions(foo_obj PUBLIC FOO=1)
```

Here a new OBJECT library `foo_obj` containing `f1.c.o` is defined. It
depends on `dep1_obj` and `dep2_obj`, this transitively pulling their
metadata. It also depends on `pthread`, which is a normal library.

`foo_obj` also has a public definition `-DFOO=1`.

A regular STATIC library `foo` is also defined. It depends on
`foo_obj`, which means that it pulls in its object files (`foo_obj`'s
object files are copied to the `.a` file without recompiling) and also
pulls it its transitive metadata, including the dependency on `pthread`
and `-DFOO=1`.

So, `foo_obj` is a collection of object files that we can use to combine
into a combined library, while `foo` is a static library, which we may
continue use as before.

Very important: note that `target_compile_definitions()` refers to
`foo_obj` instead of `foo`. I hope it is clear why. That is rather
annoying, but unfortunately there doesn't seem to be another war.

Also note that when rebuilding `hermes`, none of the many small
libraries are built, since `hermes` only depends on `hermesvm`, which
only depends on `xxx_obj` libraries. So, the build is not slower than
before.

Reviewed By: neildhar, avp

Differential Revision: D39264257

fbshipit-source-id: a972ef97f2b1b71ec4a52756c135efdd99756269