fix CommandLine handling of cl::ConsumeAfter
Summary:
ConsumeAfter is supposed to collect all extraneous positional arguments
after the last position argument, unless "--" is seen, in which case it
should collect everything after the "--".

But what it does instead is collect everything after the last positional
argument. For example in the following case:

   shermes file.js -o out

it incorrectly collects "-o" and "out".

Fix it.

Reviewed By: avp

Differential Revision: D39634641

fbshipit-source-id: e93934bcc0d68dd4edee9a27be7575d346f6ea3a