Add Unresolver pass to resolver
Summary:
Support marking identifiers which can be affected by `eval` and `with`
as "unresolvable", much like we do in Juno.

We don't yet support the `canRename` flag, because static Hermes
is as yet unconcerned with outputting JS source,
but there is a comment in the code that explains where to populate such
a flag if it exists.

Differential Revision: D40068199

fbshipit-source-id: 0c13d710adbeb1dda2f9a5dcd3f35aa485a1a516