Support named groups for static compiled RegExps
Summary: Add support for literal regexps. Add another call to parse the regex in `ESTreeIRGen::genExpression`. Use the information in the returned group name mapping to emit a `AllocObjectFromBufferInst` followed by a call to a new builtin `initRegexNamedGroups`. This builtin will take the group name mapping object and write it into the named groups object field on the `JSRegExp`.

Reviewed By: jpporto

Differential Revision: D39184317

fbshipit-source-id: e812b2001d32d0415084b89ec36943624ac61b6a