Parse named capture groups
Summary:
Support named capture groups in the regex parser. The syntax for a capture group is-

`/(?<email>.*gmail.*)/`

So more specifically, it is an open paren, question mark, and opening angle bracket, followed by the name, then the group contents. The spec for what is a valid capture group name is described here https://262.ecma-international.org/12.0/#prod-RegExpIdentifierName

NOTE- duplicate group names are disallowed by the spec but not yet handled in the parser.

Reviewed By: jpporto

Differential Revision: D38284626

fbshipit-source-id: 191685595a8d541ce36203fb715e7e9ce62a1fe2