Add new SemanticResolver (#829)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/829

Implement a new SemanticResolver which extracts all the resolution
information into `SemContext`.

This is based directly on Juno's implementation with the same
structure except for the fact that we use RAII classes instead
of the callback structure. There's nothing preventing the callback
structure from being used here, and helper methods for using callbacks
can be easily added in the future, it's just that RAII seemed fairly
straightforward and simple for this implementation.

With the exception of the Unresolver pass,
everything in the Juno implementation is contained in this diff.

The Hermes implementation also does some manipulation of the AST,
including the try/catch transform, which will be in future diffs.

Reviewed By: tmikov

Differential Revision: D39985175

fbshipit-source-id: ef99b5d8a36ee426408a213c65054362003f0995