Prevent a race condition
Summary:
When capturing JS stack traces for dApp loom, there is possibly a race condition that we try to stop the profiler from one thread, while the JS profiler from anther thread is trying to collect JS stack trace. When the app crashed, this is the crash report: P538093527

The fix here is to never collect stack trace if the profiler is already stopped.

Reviewed By: jpporto

Differential Revision: D40356058

fbshipit-source-id: ee54bf6d661a4e480c02eff45f2ea4d69c4d85e2