Explicit IR scopes for Functions
Summary:
This diff introduces a new IR type, ```ScopeDesc``` which behaves similarly to a ```VariableScope```, but without the
"restriction" that each function will have exactly one of these.

This diff introduces a new command line flag ("--enable-new-dump-format") that can be
used to inspect what the IR looks like with ScopeDescs in them. This flag is removed
once the refactor work removing VariableScopes is completed.

Reviewed By: fbmal7

Differential Revision: D38440883

fbshipit-source-id: ddec80b0dd189db4e9b13f9aae34865283596faa