Prevent a race condition
Summary:
This is a follow up to D40356058.

John pointed out that we can still have race conditions in this scenario:
T1 calls collectStackForLoom
T2 tries to disable loom profiling
T1 acquires the lock
T2 tries to acquire the lock, and is blocked
T1 checks if loom profiling is enabled
T1 releases the lock
T1 start to collect the stack
T2 acquires the lock
T2 disables profiling
T1 tries to signal the semaphore -- crash

So the fix here is to have a member variable called "collectingStack_", only ever set during collectStackForLoom(), and then when disableLoomCollection(), we will sleep on a condition variable to wait until collectingStack_ is true.

The consequence is that when loom stops, it has to wait until the last js stack collection is done. This is fine because loom thread is a background thread and it doesn't block anything.

Reviewed By: jpporto

Differential Revision: D40396011

fbshipit-source-id: 84d656b98d9b106e72c720a2d2c2e8d3fc73c6e5