Infer types for instructions that don't have them in constructors
Summary:
These are simple instructions that we can easily infer types for,
so do that.

This is a separate diff because this logic doesn't actually exist in
constructors as is.

These are HBC instructions that don't exist until after lowering
has happened, but we can support them here to avoid breaking things
if we run `TypeInference` in other places in the pipeline.

Reviewed By: tmikov

Differential Revision: D40240654

fbshipit-source-id: 5ad7eec98d7d4d1d91786f0947fc93ad2b062151