Add getInherentType method
Summary:
Add a `getInherentType` method which encodes whether an Instruction
inherently always has a type that does not depend at all on its operands.

This will eventually be used as a clear-to value in TypeInference,
and we will change most `setType` calls in Instruction constructors
to call `getInherentType` instead.

These inherent types are intended to just copy the information that
already exists in the constructors. No new typing info is added in this
diff, it's purely mechanical.

Reviewed By: tmikov

Differential Revision: D40387968

fbshipit-source-id: fc16296a970cbcc78aef3e1e492e9c1e98083ab7