Clear types and run TypeInference from scratch each time
Summary:
In order to avoid converging based on type information which is too
loose, rederive the types for each instruction from scratch whenever
TypeInference shows up in the optimization pipeline.

Now, running to a fixed point will require us to actually have useful
code in each of the `inferXXXInst` functions.

This diff contains a few important related pieces.

1. Just put `any` as the type for every instruction that we can't easily
infer the types of (this can be refined easily in the future).
2. Modify the Phi instruction inference to infer starting from NoType,
skipping any NoType Phi inputs we see (returning `true` if we see any).
Handle Phis specially outside of the ordinary dispatch mechanism.
3. Whenever we infer for a non-Phi instruction, check if any operands
are NoType first and skip the instruction because we don't have enough
information yet.
(returning `true` to indicate we have to do another iteration).
4. Start the TypeInference pass by clearing out every type that's been
inferred in the past to the instruction's inherent type.
This way, we avoid starting out with information that's too loose
and missing out on better type bounds.
5. Update a few remaining locations to not check for changes themselves.
6. Change the `inferXXXInst` instructions to simply return the `Type`
directly, and have `inferInstruction` do the work to check for changes
and set the type on the instruction.

These changes are all intertwined, and changing them independently could
result in unintended behavior, so do it all in one diff. It's a bit
noisy, but the important parts are generally in:
* `inferInstruction`
* `inferPhi`
* `clearTypesInFunction`

We also (temporarily) remove the TDZ part of the inference pass,
because having proper types means that we have to rework TDZ.
A task has been filed.

Fixing this inference algorithm had some minor impact on the
tests, notably that `bigint` doesn't show up as a type when it can't
actually be created (so there's some instructions whose number variants
can be used instead of the slow ones).

Reviewed By: tmikov

Differential Revision: D40240656

fbshipit-source-id: 88e22cb664214c54570dda5de77e65ef5e069d8e