Use inherent or trivial types in Instruction constructors
Summary:
Now that we have the `getInherentType` function, use it in any of
the instructions that have inherent types to assign the types at
instruction creation time.

Additionally, Mov and instructions like it that simply transfer a single
operand around can have their types trivially copied across.

Any and all other type inference and calls to `setType` must only occur
in the TypeInference pass.

Thus, we have created a methodology which allows us to call `setType`,
and it has been documented on `setType`.

Reviewed By: tmikov

Differential Revision: D40387933

fbshipit-source-id: ef0658e3e23dcca9d3d85b4f558786fa1e91e19e