Add a rudimentary implementation to CacheNewObject
Summary:
Add a first-attempt, fairly conservative set of rules to the
CacheNewObject optimization pass.

If a function's first uses of `this` are as the target objects
in property stores using literal keys in the entry block of the
function, and there's no `try` surrounding them, we can add
the CacheNewObject instruction. Otherwise, we bail.

The optimization only applies if there are MIN_PROPERTIES_FOR_CACHE
writes to literal `this` properties in a row.

Differential Revision: D40532426

fbshipit-source-id: 5a9a3d43cfcbe108570127ded33817f06f4d0922