Add simple type-specific operations (#843)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/843

Add checks for the types of operands to unary, binary, conditional,
and branch instructions to speed up execution.

This is a straightforward set of checks in SH instruction selection,
with some extra code and state needed because we're emitting C and we
need to make sure the parentheses line up.

This implementation is partial and there are still more specializations
which can be added without too much extra effort.

Reviewed By: tmikov

Differential Revision: D40281679

fbshipit-source-id: 48dc92b0e2afee3082f4a155d1af0c17d2fc5b8f