Don't attach docblock
Summary:
Its important to maintain the docblock position and its also annoying to defensively code to not also duplicate the docblock node when moving statements around. So since the docblock conceptually should be attached to the program lets not attached it to any nodes and always add it back to the first statement before printing.

The big downside with our current API is this will make it impossible to modify but i can do a followup diff that allows replacing the `docblock` node.

Reviewed By: bradzacher

Differential Revision: D40705705

fbshipit-source-id: fd5176991e18e16692919aadba9bd441d583f02b