IR: remove OpKind from BinaryOperatorInst/CondBranchInst
Summary:
Similarly to the previous diff, remove the special attribute denoting
the operator and just use `ValueKind`. `BinaryOperatorInst` and
`CondBranchInst` now act as superclasses of a set of instructions,
except the subclasses aren't defined, because they would be all the
same.

Instruction::getVariety() is no longer needed since there are no hidden
attributes.

Reviewed By: avp

Differential Revision: D40707385

fbshipit-source-id: e952caee9d7077b94046b99245369f55be1d2940