IR: do not subclass concrete IR types
Summary:
This is a significant refactor of the IR types, the first of several
steps. This step eliminates all subclassing of concrete IR types. Only
abstract IR types (like Literal) can be subclassed, while all concrete
classes are final.

While in some sense this is a mechanical transformation, it required
reverse engineering the intent in many places (which is one of major
problems of subclassing concrete types).

Apply the following changes:
- Remove ValueKind definitions for abstract classes.
- New macro HERMES_IR_IN_KIND() for checking a range.
- Make VariableScope abstract and introduce FuncVariableScope.
- Make Function abstract and introduce NormalFunction.
- New abstract classes BaseStoreProperty, BaseStoreOwnProperty
- New abstract class BaseLoadProperty
- New abstract class BaseCreateLexicalChildInst, BaseCreateCallableInst
- New abstract class BaseCall

Reviewed By: avp

Differential Revision: D40665135

fbshipit-source-id: 0b009eb55ad53220b4acc62e9f81ab6f18743768