Use push model for JS trace provider
Summary:
Right now the JS trace provider uses a pull model. Specifically, every 50ms, it asks the Hermes profiler to collect the stack trace from the JS thread and write them into loom entries. The problem is that this process needs to interupt the JS thread by holding a lock, and that causes race conditions with garbage collection.

This diff changes it to use pull model. Specifically, the Hermes profiler itself already have a timerLoop which proves to work well, that timerLoop collects the stack traces every N milliseconds, we can just write the stack traces into loom after that. This model makes more sense.

Here is a summary for each diff on the stack
1. Add the new API where the sampling profiler will call
2. Remove the timer loop in FBLoomJSTraceProvider, and instead set a callback for write stack to loom.
3. Let sampling profiler call the API added in step 1
4. Clean up the old code path

Reviewed By: jpporto

Differential Revision: D40944239

fbshipit-source-id: ed63264eebbed0e1cd383e84726794895367df33