add fixture+snapshot based test tooling for translation logic
Summary:
This diff adds a fixture-based test framework for the translation logic.
Instead of declaring all the tests inline in a file and relying on special utils - it relies on files being on disk and being snapshotted.

Makes it easier to add/remove snapshots and saves us having to explicitly write out the expected result using other utils.
Also allows us to easily dump out errors and the AST.

Reviewed By: pieterv

Differential Revision: D40966416

fbshipit-source-id: 1f87234b5a648fe8e308beb6fb7c86cc308c00b4