Rename JSArray::create to JSArray::createNoAllocPropStorage
Summary:
The JSArray::create overload that takes a hidden class does not allocate
the object's property storage. Turns out that trying to allocate the property
storage unconditionally leads to a non-negligible perf regression even when
no property storage is needed. Therefore, the method can't be changed. Thus,
to avoid mistakes when using the API, it is renamed to createNoAllocPropStorage.

Reviewed By: mattbfb

Differential Revision: D41092736

fbshipit-source-id: 9f8925e249860866a16a8200c6e37b287e7f66e2