Add default TZ to list of valid TZs
Summary: For some reason, Apple will return a timezone via the default timezone and then report that it is not in its list of `knownTimeZoneNames`. However, it can still be passed to the constructor of `NSTimeZone`, so for all intents and purposes it should be considered a valid time zone. Therefore, we manually add the TZ. However, because there is a chance that the default TZ can change while the program is running, we must keep track of a list of all valid timezones. This begins as just the knownTimeZoneNames. It is then potentially appended to every time the default time zone is requested, with the value of the default time zone.

Reviewed By: neildhar

Differential Revision: D41042493

fbshipit-source-id: ec72e7692acd7a0f823f70e7f9070d8d5a4c2768