Split DeleteProperty IR into strict and loose variants
Summary:
Split the DeleteProperty IR instruction into strict and loose versions.
Add a helper function to add the instruction with the strictness of the
enclosing function as the default. (presumably this will have to be
revisited if we want to remove strictness from `Function`)

This allows us to remove `isStrictMode_` in the SH backend, since the
strictness information is available based on the instruction being
emitted.

Reviewed By: tmikov

Differential Revision: D41297902

fbshipit-source-id: aa5dfce00b87f59d3f8d1986429f992c304ca4b5