Fix up ChainExpression printing
Summary:
`foo?.[0]?.bar` was getting printed as `foo?[0]?.bar`. This was caused by 2 separate issues:
- When overriding `ChainExpression` expressions we did not clear out all the properties so in the case of `OptionalMemberExpression` there were additional properties that were confusing prettier and causing it to print invalid output.
- `OptionalMemberExpression` was assigning the `object` key to `callee`. Looks like i caused this with a previous fix (oops).

Additionally when debugging i noticed replaced nodes were not getting traversed correctly so things like `parent` properties were left in. This is due to SimpleTraversal working on a clone of the AST so it would still walk the deleted nodes instead of the new ones after replacement. To fix this, i recursively call the traversal on the new nodes and skip walking of the replaced nodes.

Reviewed By: mariusschulz

Differential Revision: D41661045

fbshipit-source-id: 3b4eb8dddc19fe44e1da963e2d9e7a7b12bd2232