Delete unused parseTSFunctionTypeWithParams
Summary:
`parseTSFunctionTypeWithParams` is unused because TS doesn't allow
the same amount of code sharing that Flow does, due to the presence
of the "constructor" function types.

```
type T = new (x: number) => string;
```
is valid TS, but invalid Flow, so we need a bit of logic to deal with
that, and `parseTSFunctionTypeWithParams` wasn't called as a result.

Delete the unused function.

Reviewed By: tmikov

Differential Revision: D41736728

fbshipit-source-id: 4c4f37bc9be02b99b6ac0849022a4789f49a55a4