Disable pushscope/popscope
Summary:
Remove the pushscope/popscope implementations.

The ManagedValues set tracks external references to Hermes values. Each time a value is added, some internal criteria may decide to collect expired values within the ManagedValues collection.

PushScope/PopScope allow callers to perform a manual collection on a subset of the ManagedValues set. PushScope inserts a delimiter (after which an arbitrary number of other values are inserted), and PopScope performs a collection from the end of the set up to the previously-inserted delimiter, releasing any expired values in the subset.

This diff disables this implementation, leaving values to be collected via the normal criteria in ManagedValues. This will make it easier to refactor the ManagedValues implementation. PushScope/PopScope could be reimplemented later atop the new ManagedValues implementation.

Reviewed By: jpporto

Differential Revision: D40367066

fbshipit-source-id: 277f2e7b0f5ed16e4aa0a663d576ee7d18fbf341