Preserve values during root-counting tests
Summary:
A couple tests measure how the number of active roots changes after executing a lamba. To more accurately measure this, any newly created values should be preserved until the measurement is made, rather than destroying them in the same lambda that created them.

The previous structure produced the correct test result, but depended on knowledge of when these roots would(n't) be garbage collected. This change removes that dependency (as the collection rules will change later in this stack).

Reviewed By: neildhar

Differential Revision: D41294068

fbshipit-source-id: af141825d365e617ea9b046d282916cdf762908d