Add ability to dynamically against Hermes VM library
Summary:
Produce a dynamically linked version of `hermesvm` which exports the SH
API functions, and make it the default way to link programs produced by
`shermes`.

There were two concerns that needed to be addressed for this to work,
and one minor concern that is left unaddressed.

## Pulling in the symbols from the `hermesvm_a` static library
We need a way to make sure all the exported symbols in the `hermesvm_a`
library are pulled in to `hermesvm`. There are a few ways to do this:
1. Somehow reference the symbols in some dummy file
2. Tell the linker explicitly to pull in those symbols using the `-u`
flag, or a symbol file
3. Use `--whole-archive` or the equivalent to force the whole static
library to be pulled in

(3) has cross-platform support in recent versions of CMake, and avoids
us needing to list all of the symbols again in another place. So it
seems the least hacky and that is what is used here.

## Setting the visibility of exported functions
There are two main ways to do this:
1. Set all functions to be visible, and then provide a linker version
script or a file listing the symbols we want to export. Supplying that
script during linking will hide all the functions that are not
explicitly exported.
2. Hide all symbols by default, and use visibility attributes on the
functions that we want to export.

Doing (1) would actually also pull in the symbols, solving the previous
problem. But it has some drawbacks:

1. The symbols currently exported by the Hermes API use the second
method, so we would need to also change those.
2. It is easier to make (2) portable. The various linker flags can get
annoying to manage.

This diff uses option 2.

## Reconciling with libhermes.so
We have an existing shared library that we export called
`libhermes.so`, which also exports the actual Hermes API. For now, we
don't link against this, since the existing shermes tool directly links
against `hermesvm_a`. Eventually, we probably want to move to linking
against `libhermes.so`.

Reviewed By: tmikov

Differential Revision: D41695589

fbshipit-source-id: 49a40a6f2da293d181e2eab9d2614cb1a1a856e6