Add default TZ to list of valid TZs
Summary:
For some reason, Apple will return a timezone via the default timezone and then report that it is not in its list of `knownTimeZoneNames`. However, it can still be passed to the constructor of `NSTimeZone`, so for all intents and purposes it should be considered a valid time zone. Therefore, we manually add the TZ. However, because there is a chance that the default TZ can change while the program is running, we must keep track of a list of all valid timezones. This begins as just the knownTimeZoneNames. It is then potentially appended to every time the default time zone is requested, with the value of the default time zone.

Original Author: fbmal7@meta.com
Original Git: 1fe1ccb49bdb16ef559bf052baa42ed20b7589ca

Original Reviewed By: neildhar

Original Revision: D41042493

Reviewed By: avp

Differential Revision: D41625267

fbshipit-source-id: 0f65d3528321dac3f07d0b99d16a7aef6810b47a