use ConsecutiveStringStorage to dedup serialized literals
Summary:
Serialized literals de-duping was quadratic with the total size of the
literals because it performed a linear search when adding every literal.

We can use ConsecutiveStringStorage (with small modifications) instead
to dedup the literals with much better performance. To that end, we need
an additional pass over the LIR to collect all literals, dedup them, and
remember their dedupped offsets.

Reviewed By: tmikov

Differential Revision: D18056232

fbshipit-source-id: 14df6f5c6868339dcb11a663ab2132916bf992f1