Fix conversion of indexed to named properties in non-extensible objects
Summary:
We currently call `addOwnProperty` when converting an indexed property
to a named property. This ends up going through the extensibility check
we have for adding new properties, which means that we fail to convert
the property if the object is non-extensible.

Fix this by bypassing the check and calling `addOwnPropertyImpl`
directly.

Reviewed By: tmikov

Differential Revision: D41885823

fbshipit-source-id: 11c652fc37ba0eef23e0ffb4f8b70608a593525f