Add a options to make debugging shermes output easier
Summary:
Debugging `shermes`-generated binaries was inconvenient.

When debugging, we need the `.c` file around so that
the debugger can see it, so add an option to allow us to not
delete the temporary directory after compiling.

The debugger also needs debug info.
Add an option to pass arbitrary extra options to `cc`.
Now, passing `-Wc,-g` will allow emitting debug info for the C file
(the option is like `-Wl,` and friends in gcc/clang).

Reviewed By: tmikov

Differential Revision: D41860649

fbshipit-source-id: 18cf8b32d5b3dd619c9ec3f3097e0d789fb7cb0b