Add the TextifiedCallee debug info table
Summary:
Add a new debug info table with function names. This is the table that, if/when
present, is used by the VM to format the error message raise when a callee is
not a callable.

Also add a string table to the Hermes bytecode debug information. This table is used to
store variable names as well as the textified callees, and stores UTF8 strings that
appear in the JS source code.

The previous debug info layout had strings "inlined" in the records where
they belonged. This implementation is fine and behaved well for storing variable
names. The textified callees, on the other hand, present a fair chance for deduplication
to be effective, thus adding a table could yield some space savings. Internal experiments
showed that using the string table yielded a 5% space savings (versus adding the textified
callees without the table).

Reviewed By: avp

Differential Revision: D41744399

fbshipit-source-id: f6ad1750f193cb3ec9ef04b437077caa86329345