Add function name string literal to call instruction
Summary:
Add a new operand to CallInst (and derived classes) which will be used to propagate
the name of the function being called to the the VM. Later, if the call instruction verifies
that its operand is not a callable, this string will be used in the raised exception's
message. This new operand to CallInst is not part of the Call bytecode -- rather, a new
table will be added to the debug info mapping the bytecode addresses of call
instructions to this string.

Reviewed By: avp

Differential Revision: D40540633

fbshipit-source-id: 0337262bd00af58414b13ab3bd31c40cc53c1498