Make Runtime non-virtual
Summary:
Make Runtime non-virtual by removing `GCCallbacks` from its inheritance
hierarchy, and deleting the dummy destructor on `HandleRootOwner`.

The former is accomplished by turning the existing `GCCallbacks`
methods into non-virtual methods on `Runtime`, and introducing a new
template that wraps a `Runtime` or `DummyRuntime` to create a subclass
of `GCCallbacks` that invokes those non-virtual methods.

This implementation was chosen because:
1. It allows code sharing between `Runtime` and `DummyRuntime`.
2. It keeps the implementation of these methods in `Runtime`.
3. It makes it easy to delete `GCCallbacks` and call into `Runtime`
directly in the future if we want to, since the methods already exist
directly on `Runtime`. We would need to figure out how to handle
`DummyRuntime` when we do so, though.

Reviewed By: tmikov

Differential Revision: D42147830

fbshipit-source-id: d6c0bb3f06361802689be2d9c42e9fd959626140