Ensure lexical names in the debug info are UTF8
Summary:
The string table in the Hermes bytecode serialized format expects to hold valid UTF8 strings
at all times. This expectation is violated if any variable name in the lexical data contains invalid
UTF8 chars. This can happen as hermes utilizes a modified version of UTF8 internally. This
modified UTF8 encoding is internal to Hermes, and thus should not be serialized.

This diff modifies the lexical table generation to ensure variable names are always encoded
using valid UTF8 characters.

This diff also renames the APIs/members dealing with names in the lexical data in the debug
info so it is clearer that the payloads must be valid UTF8 strings.

Reviewed By: tmikov, mattbfb

Differential Revision: D42152144

fbshipit-source-id: e14ae308c32c539d0a1684e238c7028dd103e1b5