Fix use of Variables in freed functions in StackPromotion
Summary:
Depending on the order in which functions are destroyed by
`StackPromotion`, it is possible for a function to have an instruction
that still references a `Variable` in a function that has already been
freed. Freeing this function then causes a use-after-free.

DCE already handles this by deferring freeing the function until after
all functions have been "erased". Use the same technique here.

Interestingly, Hermes itself doesn't demonstrate this bug, because
`ScopeDesc`s are owned as a tree that is rooted in the `Module`.
But that has drawbacks, since `ScopeDesc`s will be retained long
after a function is destroyed, and may retain references to freed
functions.

Reviewed By: avp

Differential Revision: D42176893

fbshipit-source-id: 4c650d524c165016f006592772d07bfc5fe4a5bb