Add SimpleTransform API
Summary:
Create SimpleTransform API and move some shared logic from `hermes-transform` into `hermes-parser`.

The `SimpleTransform` API is designed similar to the `SimpleTraversal` API. Its basic but powerful. The `hermes-transform` API is much more user friendly and contains great tools for working with the AST but it has some limitations such as it being tricky to make deeply nested transforms. `SimpleTransform` on the other hand allows recursive transforms by modifying the AST as you traverse. This logic will be used in a followup change for `hermes-translate` but is generally useful so i have placed it here.

Reviewed By: evanyeung

Differential Revision: D41707953

fbshipit-source-id: 838b8a6ec28ce1128a2c30f872be6e918108eb8b