Suppress UBSAN float-cast-overflow in SH.cpp
Summary:
There is an instance in `SH.cpp` where we cast from double to int to
check if a value can be represented as an integer (for readability). We
have to suppress UBSAN in that instance, as we do in other places.

While here, since the attribute needs to be applied to an entire
function, hoist the function out into a helper in `Conversions.h` and
replace all places that need to suppress this particular UBSAN
complaint with the helper.

Reviewed By: tmikov

Differential Revision: D42184534

fbshipit-source-id: d044442905e657f9fcf76b1e24ad759fdbc8e02b