Use getCallsites in SimpleCallGraphProvider
Summary:
Switch `SimpleCallGraphProvider` to using `getCallsites` from `Utils`.
This unifies the codepaths so that there is only one place where we
collect callsites, which is especially useful in the construction
lowering diffs later in the stack.

It also lets `SimpleCallGraphProvider` take advantage of the more
thorough callsite checking in `getCallsites`, which handles closures
that are captured. This improves our type inference. (see the updated
tests)

It also switches `getCallsites` to taking a set instead of a vector,
which is less prone to future changes to `CallInst` that may reference
the closure multiple times.

Reviewed By: avp

Differential Revision: D42160842

fbshipit-source-id: 2beb105cb95e74054de59700c5a9f7e4a8322876