Implement CacheNewObject (#870)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/870

Add a runtime implementation for `CacheNewObject`. Some details of the
implementation:
1. An entry will be permanently set to the failed state as soon as it
fails even once. This ensures that we take the inlined fast path in all
cases except where we see repeated constructor invocations with the
same prototype chain (which should be the common case).
2. Validation is performed by walking the prototype chain and checking
for any reason setting a property on the given object may not simply
involve defining a new own property on that object. This includes the
presence of accessors with matchingnames, or the presence of a
HostObject/Proxy in the chain.
3. Any modification to the classes of objects in the prototype chain
will result in failure.
4. The cached class is constructed from the class of `thisArg`. We
require that the class of `thisArg` is a root class, since we know it
is a constructor invocation. We also require that for a given entry,
the initial class of `thisArg` is always the same.

Differential Revision: D42005920

fbshipit-source-id: 640d06bf7d2d732db862b3549ca6866a47ccd058