Disable HBC and the interpreter (#874)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/874

To allow SH to break HBC and the interpreter,
disable code generation and interpretation of the bytecode.
Mostly, this is just adding `#if 0` in a few places
to make sure things just don't build.
We also have to disable CMake building certain libraries.

There are some tests in `shermes/` that use `eval`.
They have been marked unsupported.

The InternalBytecode has also been temporarily disabled,
meaning we don't have Promise support for now.
We'll have to set it up to compile to native with `shermes`.
The `shermes/` tests that test for Promise have been disabled.

Reviewed By: neildhar

Differential Revision: D41853244

fbshipit-source-id: defaaaf360b868e845f9911e91434159bc0c3ea1