Perform TypeInference later during optimisation
Summary:
We currently run `TypeInference` before a number of other passes that
do not benefit from type information. In fact, `TypeInference` benefits
from the optimisations they make, since it is easier to infer types
when unused and unnecessary portions of the code have been removed.

To improve this, move `TypeInference` to occur later in the pipeline,
right before we run `InstSimplify` again.

Concretely, this helps the constructor optimisation later in the stack,
because the first round of `TypeInference` can observe the potentially
unreachable `return undefined` we insert at the end of a function. That
gets optimised out by `SimplifyCFG`, so running `TypeInference` after
`SimplifyCFG` allows us to infer a narrower type for the function.

Reviewed By: avp

Differential Revision: D42248067

fbshipit-source-id: 2d84c37d74e211cd58729b996971b580c121b76e