Add GetTemplateObjectInst
Summary:
Add a new IR instruction GetTemplateObjectInst which allows us to
lower it specially for SH, which can't do the same things as HBC.

The runtime behavior of the operation requires use of the template
cache, which HBC stores in the RuntimeModule.
RuntimeModule doesn't exist in SH, so we need new behavior to read
the SHUnit template cache.

This instruction is lowered into the original HermesBuiltin call
in HBC, so execution will remain unchanged there.

SH will implement a new function in the C API which will replicate this
behavior.

This diff doesn't emit the new instruction, so nothing changes yet.

Differential Revision: D42379732

fbshipit-source-id: 618a009c354d86b3efc96d5276ee3cf76c795172