TypeInference: don't default variables to "any"
Summary:
Variables with no known stores were defaulted to "any" instead of
"notype", leading to a complicated failure when there is a single store
from another variable that depends on this one.

This changes inferMemoryLocation() to behave more "naturally", by simply
generating a union of all stored types. Previously it would return `any`
if there were no stores, which is illogical, and can cause the algorithm
to fail to converge for specially constructed (but legal) IR:

```
  %2 = StoreFrameInst undefined : undefined, [i]
  %3 = StoreFrameInst 0 : number, [i]
  %6 = BranchInst %BB1
%BB1:
  %7 = LoadFrameInst [i]
  %8 = StoreFrameInst %7, [i#1]
  ...
  %15 = LoadFrameInst [i]
  %16 = UnaryOperatorInst '++', %15
  %17 = StoreFrameInst %16, [i]
  ....
  %23 = LoadFrameInst [i#1]
  %24 = StoreFrameInst %23, [i]
  %25 = StoreStackInst false : boolean, %4
  %26 = BranchInst %BB1
%BB5:
  %27 = ReturnInst undefined : undefined
```

A side effect of this change is that during the normal course of
inference, some instruction with results will temporarily have a
`notype` type. Previously we would assert against that condition
immediately, but it is just a transitional state. The algorithm will
always converge on a valid type. So, we move the assertion to the very
end.

Reviewed By: avp

Differential Revision: D42243172

fbshipit-source-id: 126f57a3ebfcd417d5d5cab950649a0dbf49200c