Add mayHaveAccessor field to ClassFlags
Summary:
Add a new field to the struct `ClassFlags`, which represents if there
is *potentially* a property accessor in the `HiddenClass`. This field
is pessimistic and can sometimes give false positives. This is due to
the fact that the field may only ever be set, but never cleared. So if
you define a getter property on an object, and then delete that
property, the field will remain set. This field will open up the door
to some potential optimizations now that we have this knowledge. It is
also used to fix a correctness bug in the next diff.

Reviewed By: neildhar

Differential Revision: D41751159

fbshipit-source-id: 5ce88c9d84c989302525df7f87a172c5510b7c51