Change printed names of IR instructions
Summary:
Annotate every IR instruction with a string used for printing. This
required some macro games, but not too much.

The printed instruction name now consists of a list of optional
prefixes, the instruction identifier, and a list of optional suffixes.
The prefixes and suffixes are always in the same order.

The currently defined prefixes:
- `js.` for a "Legacy JS" instruction.
- `hbc.` for a HBC-specific instruction

The suffixes:
- `.ns` for loose mode functionality
- `.s` for strict mode functionality
- `.r` for instructions that may recursively invoke more JS

The instruction identifiers were chosen to usually correspond to the
existing instruction kind, with some abbreviations, and start with a
lower case.

In many cases I would have preferred shorter or just different names,
but that wouldn't be practical, since it would make it very hard to
match the printed names to the instruction classes. We may consider
renaming some of the classes.

An example IR dump:
```
%BB0:
  %0 = js.loadParam (:any) %len: any
  %1 = js.lt.r (:boolean) 0: number, %0: any
  %2 = condBranch %1: boolean, %BB1, %BB2
%BB2:
  %3 = js.return (:undefined) undefined: undefined
```

Differential Revision: D42297942

fbshipit-source-id: da253e42f4ae05d91177bf5ab756e1f14b396203