Avoid widening types in TypeInference
Summary:
Ensure that TypeInference never gives worse info than it started with,
now that we're going from notype -> types in the main inference loop.

Store the pre-pass types for values when clearing them.
After running inference we can go over each value,
and check if the pre-pass type is a proper subsets of the inferred type.
If it is, reset the type to the pre-pass type.

Reviewed By: tmikov

Differential Revision: D42425123

fbshipit-source-id: f9b1d1284ac08705617b17879f0245dbd5179527