Ensure linking against the correct library
Summary:
Synthesize a library "model" in the config file and a dummy library
symbol with a name based on the model. If the client code imports this
symbol, it is guaranteed to be linking against the correct library.

For example, the symbol could be `_sh_model_cp_bd_p8_rel`. The meaning
is:
- Compress pointers
- Boxed doubles
- Non-contiguous heap (the lack of "_ch").
- 8-byte pointers (this is not strictly neccessary)
- Release build

Differential Revision: D42468377

fbshipit-source-id: e10923980bd48974cb2b5da8de234d4f54420936