Fix date-locale tests for non-US developers (#881)
Summary:
I was running the test suite (check-hermes target) on my non-american macbook and hit a test failure in the date-locale.js test as the third invocation is making the assumption that it's being run on a "en_US" machine.

Pull Request resolved: https://github.com/facebook/hermes/pull/881

Test Plan: I removed the third invocation as I believe it makes little sense to assert something which depends on system settings of the machine running the test.

Reviewed By: neildhar

Differential Revision: D42480964

Pulled By: jpporto

fbshipit-source-id: 2661f2829277d22c9940fd263ac407552cd38b4d